DF/09,16: move category detection from stage 09 to 16.
Updated:
* data samples
* run script for `ora2es` chain
