+++ Utils/Dataflow/016_task2es/task2es.py
@@ -12,6 +12,8 @@ import os
 import sys
 import traceback
 
+import re
+
 try:
     base_dir = os.path.dirname(__file__)
     dkb_dir = os.path.join(base_dir, os.pardir)
@@ -25,12 +27,93 @@ except Exception, err:
     sys.exit(1)
 
 
+def get_category(row):
+    """
+    Each task can be associated with a number of Physics Categories.
+    1) search category in hashtags list
+    2) if not found in hashtags, then search category in phys_short
+       field of tasknames
+    :param row
+    :return:
+    """
+    hashtags = row.get('hashtag_list')
+    taskname = row.get('taskname')
+    PHYS_CATEGORIES_MAP = {
+        'BPhysics': ['charmonium', 'jpsi', 'bs', 'bd', 'bminus', 'bplus',
+                     'charm', 'bottom', 'bottomonium', 'b0'],
+        'BTag': ['btagging'],
+        'Diboson': ['diboson', 'zz', 'ww', 'wz', 'wwbb', 'wwll'],
+        'DrellYan': ['drellyan', 'dy'],
+        'Exotic': ['exotic', 'monojet', 'blackhole', 'technicolor',
+                   'randallsundrum', 'wprime', 'zprime', 'magneticmonopole',
+                   'extradimensions', 'warpeded', 'randallsundrum',
+                   'contactinteraction', 'seesaw'],
+        'GammaJets': ['photon', 'diphoton'],
+        'Higgs': ['whiggs', 'zhiggs', 'mh125', 'higgs', 'vbf', 'smhiggs',
+                  'bsmhiggs', 'chargedhiggs'],
+        'Minbias': ['minbias'],
+        'Multijet': ['dijet', 'multijet', 'qcd'],
+        'Performance': ['performance'],
+        'SingleParticle': ['singleparticle'],
+        'SingleTop': ['singletop'],
+        'SUSY': ['bino', 'susy', 'pmssm', 'leptosusy', 'rpv', 'mssm'],
+        'Triboson': ['triplegaugecoupling', 'triboson', 'zzw', 'www'],
+        'TTbar': ['ttbar'],
+        'TTbarX': ['ttw', 'ttz', 'ttv', 'ttvv', '4top', 'ttww'],
+        'Upgrade': ['upgrad'],
+        'Wjets': ['w'],
+        'Zjets': ['z']}
+    match = {}
+    categories = []
+    for phys_category in PHYS_CATEGORIES_MAP:
+        current_map = [x.strip(' ').lower()
+                       for x in PHYS_CATEGORIES_MAP[phys_category]]
+        if hashtags is not None:
+            match[phys_category] = len([x for x in hashtags.lower().split(',')
+                                        if x.strip(' ') in current_map])
+    categories = [cat for cat in match if match[cat] > 0]
+    if not categories and taskname:
+        phys_short = taskname.split('.')[2].lower()
+        if re.search('singletop', phys_short) is not None:
+            categories.append("SingleTop")
+        if re.search('ttbar', phys_short) is not None:
+            categories.append("TTbar")
+        if re.search('jets', phys_short) is not None:
+            categories.append("Multijet")
+        if re.search('h125', phys_short) is not None:
+            categories.append("Higgs")
+        if re.search('ttbb', phys_short) is not None:
+            categories.append("TTbarX")
+        if re.search('ttgamma', phys_short) is not None:
+            categories.append("TTbarX")
+        if re.search('_tt_', phys_short) is not None:
+            categories.append("TTbar")
+        if re.search('upsilon', phys_short) is not None:
+            categories.append("BPhysics")
+        if re.search('tanb', phys_short) is not None:
+            categories.append("SUSY")
+        if re.search('4topci', phys_short) is not None:
+            categories.append("Exotic")
+        if re.search('xhh', phys_short) is not None:
+            categories.append("Higgs")
+        if re.search('3top', phys_short) is not None:
+            categories.append("TTbarX")
+        if re.search('_wt', phys_short) is not None:
+            categories.append("SingleTop")
+        if re.search('_wwbb', phys_short) is not None:
+            categories.append("SingleTop")
+        if re.search('_wenu_', phys_short) is not None:
+            categories.append("Wjets")
+    if not categories:
+        categories = ["Uncategorized"]
+    return categories
+
+
 def process(stage, message):
     """ Single message processing. """
     data = message.content()
-    # Processing machinery
-    out_data = {"key": "value"}
-    out_message = JSONMessage(out_data)
+    data['phys_category'] = get_category(data)
+    out_message = JSONMessage(data)
     stage.output(out_message)
     return True
